Address PR review: use rapidfuzz, dynamic remotes, extract utils
- Replace hand-rolled Levenshtein with rapidfuzz library
- Use actual git remotes (from `git remote`) instead of hardcoded
  prefixes like "origin"/"upstream" for stripping remote prefixes
- Move matching functions to forestui/utils.py, keeping only widget
  classes in branch_search.py
- Thread remotes list from git service through app -> modals -> widgets
- Update _compute_default_base_branch to iterate actual remotes
